改成long，以支持更大的数字，但是最大也只能支持19位的数字 long.maxValue: 9223372036854775807

+++ study-codes/algorithm-demo/src/main/java/beinet/cn/algorithmdemo/controller/dto/DivisorResult.java
@@ -19,5 +19,5 @@ public class DivisorResult {
     @ApiModelProperty(value = "原数=约数的乘积算式")
     private String mul;
     @ApiModelProperty(value = "属于质数的约数列表")
-    private Set<Integer> prime;
+    private Set<Long> prime;
 }
